requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
configobj==5.0.6
cycler==0.11.0
dulwich==0.20.43
everett==3.0.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
grpcio==1.46.3
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==4.11.4
joblib==1.1.0
jsonschema==4.6.0
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
networkx==2.8.4
numpy==1.22.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyDeprecate==0.3.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.6.4
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
semantic-version==2.10.0
Shapely==1.8.2
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.8.2+cu102
torchmetrics==0.9.1
torchvision==0.9.2+cu102
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.2
wrapt==1.14.1
wurlitzer==3.0.2
yarl==1.7.2
zipp==3.8.0
